feat: rename nicknamebottom
